修复参数类型

+++ core/os/file_access_http.h
@@ -113,7 +113,7 @@ public:
 	virtual String get_path() const; /// returns the path for the current open file
 	virtual String get_path_absolute() const; /// returns the absolute path for the current open file
 
-	virtual void seek(size_t p_position); ///< seek to a given position
+	virtual void seek(uint64_t p_position); ///< seek to a given position
 	virtual void seek_end(int64_t p_position = 0); ///< seek from the end of file
 	virtual size_t get_position() const; ///< get position in the file
 	virtual size_t get_len() const; ///< get size of the file
@@ -121,7 +121,8 @@ public:
 	virtual bool eof_reached() const; ///< reading passed EOF
 
 	virtual uint8_t get_8() const; ///< get a byte
-	virtual int get_buffer(uint8_t *p_dst, int p_length) const ;
+
+	virtual uint64_t get_buffer(uint8_t *p_dst, uint64_t p_length) const ;
 
 	virtual Error get_error() const; ///< get last error
